// app/models/Text.ts
type Text = {
    introText: string | null;
    yardUtilitiesText: string | null;
    ourYardText: string | null;
    missionVisionText: string | null;
    visionText: string | null;
    valuesText: string | null;
    missionText: string | null;
    OurTechnologyIntroTexto: string | null;
    introTextOurServices: string | null;
    services: string[];
    services2: string[];
    installationUtilities: string | null;
    productionEquipmentIntro: string | null;
    logisticEquipmentIntro: string | null;
    QcEquipmentTitle: string | null;
    QcEquipmentIntro: string | null;
    qualityHealthSafetyAndEnvironmentalcareIntroText: string | null;
    ManpowerAndTrainingText: string | null;
    ShallowWaterFieldsTitle: string | null;
    ShallowWaterFieldsIntro: string | null;
    DeepWaterFieldsTitle: string | null;
    DeepWaterFieldsIntro: string | null;
};


export default Text